# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Device Management & Measurement API",
    "description": "API to manage customers, devices, and ingest/retrieve measurements.",
    "version": "0.1.0"
  },
  "paths": {
    "/customers": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Create a customer",
        "description": "Register a new customer.\n\n- Returns **201 Created** with the created customer.\n- Returns **409 Conflict** if the email is already registered.\n- Returns **422 Unprocessable Entity** if validation fails.",
        "operationId": "create_customer_customers_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "example": {
                  "detail": "A resource with the same unique value already exists"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "List customers",
        "description": "List customers with optional filters and cursor-based pagination.\n\n- Returns **200** with a paginated list of customers and a cursor for the next page.\n- Filters combine with AND logic.\n- Returns **422** if filter values fail validation.",
        "operationId": "list_customers_customers_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "email"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exact email address",
              "title": "Email"
            },
            "description": "Filter by exact email address"
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ISO 3166-1 alpha-2 country code",
              "title": "Country"
            },
            "description": "Filter by ISO 3166-1 alpha-2 country code"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ISO 639-1 language code",
              "title": "Language"
            },
            "description": "Filter by ISO 639-1 language code"
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by partial name match (case-insensitive)",
              "title": "Name"
            },
            "description": "Filter by partial name match (case-insensitive)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(created_at|name|email)$",
              "description": "Field to sort by",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "Field to sort by"
          },
          {
            "name": "sort_dir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort direction",
              "default": "asc",
              "title": "Sort Dir"
            },
            "description": "Sort direction"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum results per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum results per page"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination cursor from previous response",
              "title": "Cursor"
            },
            "description": "Opaque pagination cursor from previous response"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCustomersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customer_id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Get a customer",
        "description": "Retrieve a customer by ID.\n\n- Returns **404 Not Found** if the customer does not exist.\n- Returns **422 Unprocessable Entity** if `customer_id` is not a valid UUID.",
        "operationId": "get_customer_customers__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Resource with the given id was not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/devices/{device_id}/customer": {
      "put": {
        "tags": [
          "Devices"
        ],
        "summary": "Assign device to a customer",
        "description": "Assign a device to a customer, optionally setting a timezone.\n\n- Idempotent: re-assigning to the same or a different customer is always accepted.\n- Returns **404** if the device or customer does not exist.\n- Returns **422** if payload validation fails.",
        "operationId": "assign_device_customer_devices__device_id__customer_put",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Resource with the given id was not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Devices"
        ],
        "summary": "Unassign device from its customer",
        "description": "Remove the customer assignment from a device and clear its timezone.\n\n- Idempotent: unassigning an already-unassigned device is a successful no-op.\n- Returns **404** if the device does not exist.",
        "operationId": "unassign_device_customer_devices__device_id__customer_delete",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Resource with the given id was not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique customer identifier"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Customer name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Customer email address"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "ISO 639-1 2-letter language code"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country",
            "description": "ISO 3166-1 alpha-2 country code"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone",
            "description": "IANA Time Zone identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "UTC timestamp when the customer was created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "created_at"
        ],
        "title": "CustomerResponse",
        "description": "Schema for customer response."
      },
...
      "DeviceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique device identifier"
          },
          "serial_number": {
            "type": "string",
            "title": "Serial Number",
            "description": "Unique device serial number"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id",
            "description": "ID of the assigned customer; null if unassigned"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Device lifecycle status (e.g. 'unassigned', 'assigned')"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the device is active and can ingest measurements"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone",
            "description": "IANA Time Zone identifier assigned to this device"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "UTC timestamp when the device was registered"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "UTC timestamp of the last update"
          }
        },
        "type": "object",
        "required": [
          "id",
          "serial_number",
          "customer_id",
          "status",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "DeviceResponse",
        "description": "Schema for device response."
      },
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable description of the error"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "description": "Body returned by all domain-level error handlers (404, 409, 500)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedCustomersResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CustomerResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Customers on this page"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque cursor to pass as `cursor` to retrieve the next page; null if no more pages"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether additional pages exist"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "title": "PaginatedCustomersResponse",
        "description": "Paginated list of customers."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationErrorItem": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc",
            "description": "Path to the failing field, e.g. ['body', 'email']"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Human-readable reason the validation failed"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Machine-readable error type identifier"
          },
          "input": {
            "title": "Input",
            "description": "The value that was rejected"
          },
          "ctx": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctx",
            "description": "Extra context from the validator"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationErrorItem",
        "description": "A single Pydantic / FastAPI validation failure."
      },
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "description": "Body returned on 422 Unprocessable Content."
      }
    }
  }
}